+++ gripper/install the gripper
@@ -1,0 +1,12 @@
+  
+// Installing required package
+
+cd ~/catkin_ws/src
+git clone https://github.com/ros-drivers/rosserial.git
+cd ~/catkin_ws
+catkin_make
+
+//Next, run the rosserial client application that forwards your Arduino messages to the rest of ROS. Make sure to use the correct serial port
+
+rosrun rosserial_python serial_node.py /dev/ttyUSB0
+
